Adding tests and program skeleton.
The tests that I have previously completed for Database connection and API testing are included here.

Also I have provided the classes in the "rockmatching" folder. This is an Eclipse project that is set up for Maven web development, and includes all the nevessary dependencies for api and database integration.